Add packaging metadata
